refactor: major BookingService restructuring and separation of concerns
Breaking Changes:
- Removed ExpectedPrice field from CreateBookingDto
- Backend now autonomously calculates booking prices without client verification

New Services Created:
- BookingValidationService: Handles all validation logic (pickup time, route verification, Arlanda requirements, price calculation)
- BookingFactoryService: Responsible for entity creation, coordinate list building, and price calculation request building
- BookingMapperService: Handles all entity-to-DTO mapping for bookings

Service Improvements:
- BookingService: Reduced from ~500 lines to ~300 lines, now acts as coordinator

Structural Changes:
- Created DTOs/ValidationResults folder for internal validation result classes
  - ValidationResult.cs
  - RouteValidationResult.cs
  - PriceValidationResult.cs
- All booking-related services follow consistent "Service" naming suffix
- Updated DependencyInjections.cs to register all new services

Benefits:
- Improved separation of concerns (Single Responsibility Principle)
- Enhanced testability with smaller, focused services
- Better code maintainability and readability
- Easier to extend with new validation rules or mapping logic
- Clear service boundaries and responsibilities

Updated Files:
- Services/Implementations/BookingService.cs (refactored)
- Services/Implementations/BookingValidationService.cs (new)
- Services/Implementations/BookingFactoryService.cs (new)
- Services/Implementations/BookingMapperService.cs (new)
- Services/Implementations/EmailService.cs (enhanced)
- Services/Interfaces/IBookingValidationService.cs (new)
- Services/Interfaces/IBookingFactoryService.cs (new)
- Services/Interfaces/IBookingMapperService.cs (new)
- Services/Interfaces/IEmailService.cs (updated)
- DTOs/BookingDTOs/CreateBookingDto.cs (removed ExpectedPrice)
- DTOs/ValidationResults/*.cs (new)
- Configurations/DependencyInjections.cs (updated registrations)
- Validators/CreateBookingDTOValidator.cs (removed ExpectedPrice validation)
