+++ Configurations/DependencyInjections.cs
@@ -4,7 +4,9 @@ using PegasusBackend.DTOs.EmailDTO;
 using PegasusBackend.Repositorys.Implementations;
 using PegasusBackend.Repositorys.Interfaces;
 using PegasusBackend.Services.Implementations;
+using PegasusBackend.Services.Implementations.BookingServices;
 using PegasusBackend.Services.Interfaces;
+using PegasusBackend.Services.Interfaces.BookingInterfaces;
 
 namespace PegasusBackend.Configurations
 {
@@ -27,11 +29,15 @@ namespace PegasusBackend.Configurations
             services.AddScoped<IMapService, MapService>();
             services.AddScoped<IBookingService, BookingService>();
 
+            // Booking-related services
+            services.AddScoped<IBookingValidationService, BookingValidationService>();
+            services.AddScoped<IBookingFactoryService, BookingFactoryService>();
+            services.AddScoped<IBookingMapperService, BookingMapperService>();
+
             // Email Configuration & Service
             var emailConfig = configuration.GetSection("EmailConfig").Get<EmailConfig>();
             services.AddSingleton(emailConfig!);
             services.AddScoped<IEmailService, EmailService>();
-
 
             return services;
         }
