+++ Services/Interfaces/BookingInterfaces/IBookingMapperService.cs
@@ -1,0 +1,11 @@
+﻿using PegasusBackend.DTOs.BookingDTOs;
+using PegasusBackend.Models;
+
+namespace PegasusBackend.Services.Interfaces.BookingInterfaces
+{
+    public interface IBookingMapperService
+    {
+        BookingResponseDto MapToResponseDTO(Bookings booking);
+        List<BookingResponseDto> MapToResponseDTOs(List<Bookings> bookings);
+    }
+}
